ports in kubernetes are int32 now

+++ pkg/transformers/kubernetes/generate.go
@@ -181,7 +181,7 @@ func Generate(vals *types.CmdValues) (*kapi.List, error) {
 			for _, s := range v.Ports {
 				container, _ := extractFirstPorts(s)
 				if port, err := strconv.Atoi(container); err == nil {
-					c.Ports = append(c.Ports, kapi.ContainerPort{ContainerPort: port})
+					c.Ports = append(c.Ports, kapi.ContainerPort{ContainerPort: int32(port)})
 				}
 			}
 			c.Args = v.Command.Slice()
